Removed partial guild classes and merged with existing base engine classes.
