+++ Intersect (Core)/Intersect (Core).csproj
@@ -278,7 +278,6 @@
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Factories\FactoryRegistry.cs" />
     <Compile Include="Factories\IFactory.cs" />
-    <Compile Include="GameObjects\Events\Commands\GuildEventCommands.cs" />
     <Compile Include="GameObjects\Events\VariableMod.cs" />
     <Compile Include="GameObjects\Switches and Variables\VariableValue.cs" />
     <Compile Include="GlobalSuppressions.cs" />
