--- Intersect.Client/Entities/Player.Guilds.cs
@@ -1,87 +1,0 @@
-﻿using Intersect.Client.Core;
-using Intersect.Client.General;
-using Intersect.Config.Guilds;
-using Intersect.Enums;
-using Intersect.Network.Packets.Server;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Intersect.Client.Entities
-{
-    public partial class Player
-    {
-        public string Guild;
-
-        public int Rank;
-
-        public bool InGuild => !string.IsNullOrWhiteSpace(Guild);
-
-        public GuildRank GuildRank => InGuild ? Options.Instance.Guild.Ranks[Math.Max(0, Math.Min(this.Rank, Options.Instance.Guild.Ranks.Length - 1))] : null;
-
-        public virtual void DrawGuildName(Color textColor, Color borderColor = null, Color backgroundColor = null)
-        {
-            if (HideName || Guild == null || Guild.Trim().Length == 0 || !Options.Instance.Guild.ShowGuildNameTagsOverMembers)
-            {
-                return;
-            }
-
-            if (borderColor == null)
-            {
-                borderColor = Color.Transparent;
-            }
-
-            if (backgroundColor == null)
-            {
-                backgroundColor = Color.Transparent;
-            }
-
-            //Check for stealth amoungst status effects.
-            for (var n = 0; n < Status.Count; n++)
-            {
-                //If unit is stealthed, don't render unless the entity is the player.
-                if (Status[n].Type == StatusTypes.Stealth)
-                {
-                    if (this != Globals.Me && !(this is Player player && Globals.Me.IsInMyParty(player)))
-                    {
-                        return;
-                    }
-                }
-            }
-
-            var map = MapInstance;
-            if (map == null)
-            {
-                return;
-            }
-
-            var textSize = Graphics.Renderer.MeasureText(Guild, Graphics.EntityNameFont, 1);
-
-            var x = (int)Math.Ceiling(GetCenterPos().X);
-            var y = GetLabelLocation(LabelType.Guild);
-
-            if (backgroundColor != Color.Transparent)
-            {
-                Graphics.DrawGameTexture(
-                    Graphics.Renderer.GetWhiteTexture(), new Framework.GenericClasses.FloatRect(0, 0, 1, 1),
-                    new Framework.GenericClasses.FloatRect(x - textSize.X / 2f - 4, y, textSize.X + 8, textSize.Y), backgroundColor
-                );
-            }
-
-            Graphics.Renderer.DrawString(
-                Guild, Graphics.EntityNameFont, (int)(x - (int)Math.Ceiling(textSize.X / 2f)), (int)y, 1,
-                Color.FromArgb(textColor.ToArgb()), true, null, Color.FromArgb(borderColor.ToArgb())
-            );
-        }
-
-
-        /// <summary>
-        /// Contains a record of all members of this player's guild.
-        /// </summary>
-        public GuildMember[] GuildMembers = new GuildMember[0];
-
-
-    }
-}
