+++ Intersect.Client/Entities/Player.cs
@@ -22,6 +22,7 @@ using Intersect.Client.Framework.GenericClasses;
 using Intersect.Utilities;
 using Intersect.Client.Items;
 using Intersect.Client.Interface.Game.Chat;
+using Intersect.Config.Guilds;
 
 namespace Intersect.Client.Entities
 {
@@ -78,6 +79,31 @@ namespace Intersect.Client.Entities
 
         private Dictionary<int, long> mLastHotbarUseTime = new Dictionary<int, long>();
         private int mHotbarUseDelay = 150;
+
+        /// <summary>
+        /// Name of our guild if we are in one.
+        /// </summary>
+        public string Guild;
+
+        /// <summary>
+        /// Index of our rank where 0 is the leader
+        /// </summary>
+        public int Rank;
+
+        /// <summary>
+        /// Returns whether or not we are in a guild by checking to see if we are assigned a guild name
+        /// </summary>
+        public bool InGuild => !string.IsNullOrWhiteSpace(Guild);
+
+        /// <summary>
+        /// Obtains our rank and permissions from the game config
+        /// </summary>
+        public GuildRank GuildRank => InGuild ? Options.Instance.Guild.Ranks[Math.Max(0, Math.Min(this.Rank, Options.Instance.Guild.Ranks.Length - 1))] : null;
+
+        /// <summary>
+        /// Contains a record of all members of this player's guild.
+        /// </summary>
+        public GuildMember[] GuildMembers = new GuildMember[0];
 
         public Player(Guid id, PlayerEntityPacket packet) : base(id, packet)
         {
@@ -1862,6 +1888,61 @@ namespace Intersect.Client.Entities
             DrawLabels(HeaderLabel.Text, 0, HeaderLabel.Color, textColor, borderColor, backgroundColor);
             DrawLabels(FooterLabel.Text, 1, FooterLabel.Color, textColor, borderColor, backgroundColor);
             DrawGuildName(textColor, borderColor, backgroundColor);
+        }
+
+        public virtual void DrawGuildName(Color textColor, Color borderColor = null, Color backgroundColor = null)
+        {
+            if (HideName || Guild == null || Guild.Trim().Length == 0 || !Options.Instance.Guild.ShowGuildNameTagsOverMembers)
+            {
+                return;
+            }
+
+            if (borderColor == null)
+            {
+                borderColor = Color.Transparent;
+            }
+
+            if (backgroundColor == null)
+            {
+                backgroundColor = Color.Transparent;
+            }
+
+            //Check for stealth amoungst status effects.
+            for (var n = 0; n < Status.Count; n++)
+            {
+                //If unit is stealthed, don't render unless the entity is the player.
+                if (Status[n].Type == StatusTypes.Stealth)
+                {
+                    if (this != Globals.Me && !(this is Player player && Globals.Me.IsInMyParty(player)))
+                    {
+                        return;
+                    }
+                }
+            }
+
+            var map = MapInstance;
+            if (map == null)
+            {
+                return;
+            }
+
+            var textSize = Graphics.Renderer.MeasureText(Guild, Graphics.EntityNameFont, 1);
+
+            var x = (int)Math.Ceiling(GetCenterPos().X);
+            var y = GetLabelLocation(LabelType.Guild);
+
+            if (backgroundColor != Color.Transparent)
+            {
+                Graphics.DrawGameTexture(
+                    Graphics.Renderer.GetWhiteTexture(), new Framework.GenericClasses.FloatRect(0, 0, 1, 1),
+                    new Framework.GenericClasses.FloatRect(x - textSize.X / 2f - 4, y, textSize.X + 8, textSize.Y), backgroundColor
+                );
+            }
+
+            Graphics.Renderer.DrawString(
+                Guild, Graphics.EntityNameFont, (int)(x - (int)Math.Ceiling(textSize.X / 2f)), (int)y, 1,
+                Color.FromArgb(textColor.ToArgb()), true, null, Color.FromArgb(borderColor.ToArgb())
+            );
         }
 
         public void DrawTargets()
