+++ Intersect.Editor/Intersect.Editor.csproj
@@ -131,7 +131,6 @@
       <DependentUpon>frmUpdate.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\Helpers\GridHelper.cs" />
-    <Compile Include="Localization\Strings.Guilds.cs" />
     <Compile Include="Maps\MapGridItem.cs" />
     <Compile Include="Maps\MapSaveState.cs" />
     <Compile Include="Forms\Controls\SearchableDarkTreeView.cs">
