+++ Intersect.Server/Intersect.Server.csproj
@@ -481,7 +481,6 @@
     <Compile Include="Entities\Combat\Buff.cs" />
     <Compile Include="Entities\Combat\Dash.cs" />
     <Compile Include="Entities\Combat\DoT.cs" />
-    <Compile Include="Entities\Events\CommandProcessing.Guilds.cs" />
     <Compile Include="Entities\Events\ConditionHandlerRegistry.cs" />
     <Compile Include="Entities\Events\Conditions.cs" />
     <Compile Include="Entities\Events\VariableCheckHandlerRegistry.cs" />
@@ -494,7 +493,6 @@
     <Compile Include="Database\PlayerData\User.cs" />
     <Compile Include="Database\PlayerData\PlayerContext.cs" />
     <Compile Include="Entities\Events\Event.cs" />
-    <Compile Include="Entities\Player.Guilds.cs" />
     <Compile Include="Entities\Projectile.cs" />
     <Compile Include="Entities\ProjectileSpawn.cs" />
     <Compile Include="Entities\Resource.cs" />
@@ -509,7 +507,6 @@
     <Compile Include="General\Time.cs" />
     <Compile Include="Localization\CommandsNamespace.cs" />
     <Compile Include="Localization\Strings.cs" />
-    <Compile Include="Localization\Strings.Guilds.cs" />
     <Compile Include="Maps\MapActionMessages.cs" />
     <Compile Include="Maps\MapAnimations.cs" />
     <Compile Include="Maps\MapEntityMovements.cs" />
@@ -581,7 +578,6 @@
     <Compile Include="Networking\Helpers\NetDebug.cs" />
     <Compile Include="Networking\PacketHandler.cs" />
     <Compile Include="General\Globals.cs" />
-    <Compile Include="Networking\PacketHandler.Guilds.cs" />
     <Compile Include="Networking\PacketSender.cs" />
     <Compile Include="Entities\Player.cs" />
     <Compile Include="Entities\Entity.cs" />
@@ -589,7 +585,6 @@
     <Compile Include="Maps\MapGrid.cs" />
     <Compile Include="Networking\Helpers\PortChecker.cs" />
     <Compile Include="Networking\Helpers\UPnP.cs" />
-    <Compile Include="Networking\PacketSender.Guilds.cs" />
     <Compile Include="Notifications\Notification.cs" />
     <Compile Include="Notifications\ResetPassword.cs" />
     <Compile Include="Networking\Lidgren\ServerNetwork.cs" />
