+++ Intersect.Server/Networking/PacketSender.cs
@@ -2001,6 +2001,46 @@ namespace Intersect.Server.Networking
             player.SendPacket(new TargetOverridePacket(targetId));
         }
 
+        //GuildMsgPacket
+        public static void SendGuildMsg(Player player, string message, Color clr, string target = "")
+        {
+            foreach (var p in player.Guild.FindOnlineMembers())
+            {
+                if (p != null)
+                {
+                    SendChatMsg(p, message, ChatMessageType.Guild, clr, target);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Send a player their guild member list.
+        /// </summary>
+        /// <param name="player"></param>
+        public static void SendGuild(Player player)
+        {
+            if (player == null || player.Guild == null)
+            {
+                return;
+            }
+
+            var members = player.Guild.Members.Values.ToArray();
+            var onlineMembers = player.Guild.FindOnlineMembers();
+
+            foreach (var member in members)
+            {
+                member.Online = onlineMembers.Any(m => m.Id == member.Id);
+            }
+
+            player.SendPacket(new GuildPacket(members));
+        }
+
+        //GuildRequestPacket
+        public static void SendGuildInvite(Player player, Player from)
+        {
+            player.SendPacket(new GuildInvitePacket(from.Name, from.Guild.Name));
+        }
+
         public static void SendDataToMap(Guid mapId, IPacket packet, Player except = null, TransmissionMode mode = TransmissionMode.All)
         {
             if (!MapInstance.Lookup.Keys.Contains(mapId))
